Auth Refactor
- Changed storage reliance on pinia:

Pinia is no longer required to use the auth module, but can still be used. By default it is disabled and instead nuxt's useState will be used instead.

- Moved login watch to plugin:

This is still in testing however I found that watching loggedIn works better when it's in a plugin.

- Consolidated store/storage options:

localStorage, sessionStorage, cookie, and pinia have all been moved under on property called 'stores'. Theves also been shortened in name so localStorage is 'local', cookie is 'cookie', and sessionStorage is 'session'.

The strategy cookie, will no longer be present if you're not logged in.

The syncUniversal/setUniversal now accept a third object parameter that dictates whether or not you want to exclude a store (the store wont be added in any case if its disabled in your options). In the case for cookies you can either use a boolean or an object which contains the cookie's options.