v3.0.0 (#79)
* Auth Refactor

- Changed storage reliance on pinia:

Pinia is no longer required to use the auth module, but can still be used. By default it is disabled and instead nuxt's useState will be used instead.

- Moved login watch to plugin:

This is still in testing however I found that watching loggedIn works better when it's in a plugin.

- Consolidated store/storage options:

localStorage, sessionStorage, cookie, and pinia have all been moved under on property called 'stores'. Theves also been shortened in name so localStorage is 'local', cookie is 'cookie', and sessionStorage is 'session'.

The strategy cookie, will no longer be present if you're not logged in.

The syncUniversal/setUniversal now accept a third object parameter that dictates whether or not you want to exclude a store (the store wont be added in any case if its disabled in your options). In the case for cookies you can either use a boolean or an object which contains the cookie's options.

* Cookie Scheme Refactor

 - The Cookie scheme has been added back under the local scheme. It still retains the same functionality if you're not using the token property.

-  The Laravel Sanctum provider will use the token method, if you'd like to use the  SPA method set token.type to false.

 - Documentation stackblitz has been update to reflect changes: https://stackblitz.com/edit/github-nufjhw?file=README.md

* Improve type support

- When configuring the auth options the strategies would have no type hinting. This update aims to fix this.

* preparation: v3.0.0 Release